Jurgen

+++ src/services/word.service.ts
@@ -1,15 +1,18 @@
 import { CreateWordDto, UpdateWordDto, createWordSchema, updateWordSchema } from '../dto/word';
 import { Word } from '../entities';
 import { mapJoiErrors } from '../middlewares/validation-error.middleware';
+import { WordCategoryRepository } from '../repositories/word-category.repository';
 import { WordRepository } from '../repositories/word.repository';
-import { WORD_ALREADY_EXISTS, WORD_NOT_FOUND } from '../utilities/messages.utility';
+import { WORD_ALREADY_EXISTS, WORD_ALREADY_USED_IN_CATEGORY, WORD_NOT_FOUND } from '../utilities/messages.utility';
 
 
 export class WordService {
     private wordRepository: WordRepository;
+    private wordCategoryRepository: WordCategoryRepository;
     
     constructor() {
-        this.wordRepository = new WordRepository();
+        this.wordRepository         = new WordRepository();
+        this.wordCategoryRepository = new WordCategoryRepository();
     }
 
     async getAllWords(): Promise<Word[]> {
@@ -46,8 +49,10 @@ export class WordService {
 
     async deleteWord(id: number): Promise<void> {
         const responseById = await this.wordRepository.findWordById(id);
+        const responseCategoryByIdWord = await this.wordCategoryRepository.findCategoryByIdWord(id);
 
         if (!responseById) throw new Error(WORD_NOT_FOUND);
+        if(responseCategoryByIdWord.length > 0) throw new Error(WORD_ALREADY_USED_IN_CATEGORY);
 
         this.wordRepository.deleteWord(id);
     }
